express08

+++ userHandlers.js
@@ -1,5 +1,24 @@
 const database = require("./database");
 
+const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
+  const { email } = req.body;
+
+  database
+    .query("select id, firstname, lastname, email, city, language, hashedPassword from users where email = ?", [email])
+    .then(([users]) => {
+      if (users[0] != null) {
+        req.user = users[0];
+        next();
+      } else {
+        res.sendStatus(401);
+      }
+    })
+    .catch((err) => {
+      console.error(err);
+      res.status(500).send("Error retrieving data from database");
+    });
+};
+
 const getUsers = (req, res) => {
   const initialSql = "select id, firstname, lastname, email, city, language from users";
   const where = [];
@@ -77,6 +96,10 @@ const updateUser = (req, res) => {
   const id = parseInt(req.params.id);
   const { firstname, lastname, email, city, language, hashedPassword } = req.body;
 
+  if (req.payload.sub !== id) {
+    return res.status(403).send("Forbidden");
+  }
+
   database
     .query(
       "update users set firstname = ?, lastname = ?, email = ?, city = ?, language = ?, hashedPassword = ? where id = ?",
@@ -98,6 +121,10 @@ const updateUser = (req, res) => {
 const deleteUser = (req, res) => {
   const id = parseInt(req.params.id);
 
+  if (req.payload.sub !== id) {
+    return res.status(403).send("Forbidden");
+  }
+
   database
     .query("delete from users where id = ?", [id])
     .then(([result]) => {
@@ -114,6 +141,7 @@ const deleteUser = (req, res) => {
 };
 
 module.exports = {
+  getUserByEmailWithPasswordAndPassToNext,
   getUsers,
   getUserById,
   postUser,
